extracted execute_dry_run fn
